+++ src/components/kytos/inputs/Select.vue
@@ -1,11 +1,14 @@
 <template>
    <label class="k-select no-compact">
-    <div class="k-select__title">
+    <div v-if="title" class="k-select__title">
       <icon v-if="icon && iconName"  :icon="iconName" data-test="main-icon"></icon>
       {{title}}
     </div>
+    <div class="input_filter" v-if="enable_filter">
+      <input v-model="filter_text"></input>
+    </div>
     <select class="k-select__select" v-model="selected" @change="emitEvent" multiple data-test="main-select">
-      <option v-for="item in options" :value="item.value" :key="item.value">
+      <option :class="isAMatch(item)" v-for="item in options" :value="item.value" :key="item.value">
         {{item.description}} 
       </option>
     </select>
@@ -49,11 +52,21 @@ export default {
     action: {
       type: Function,
       default: function (value) { return }
+    },
+    title: {
+      type: String,
+      required: false,
+    },
+    enable_filter: {
+      type: Boolean,
+      default: false,
+      required: false
     }
   },
   data () {
     return {
-      selected: []
+      selected: [],
+      filter_text: "",
     }
   },
   methods: {
@@ -63,7 +76,16 @@ export default {
     },
     clear () {
       this.selected = [];
-    }
+    },
+    isAMatch (option) {
+      if (this.filter_text) {
+        if (option.description.toUpperCase().includes(this.filter_text.toUpperCase())) {
+          return "matched";
+        }
+        return "not_matched";
+      }
+      return "no_filter";
+    },
   },
   beforeMount () {
     // Initialize the selected values with the v-model:value property
@@ -98,6 +120,12 @@ export default {
  font-size: 0.8em
  color: dark-theme-variables.$fill-text
 
+.k-select .input_filter input
+ background: lightgray
+ border-radius: 3px
+ font-size: 0.9em
+ width: 99.5%
+
  &:hover svg
   fill: dark-theme-variables.$fill-icon-h
   color: dark-theme-variables.$fill-icon-h
@@ -106,12 +134,6 @@ export default {
    padding-bottom: 2px
    padding-left: 3px
    padding-top: 0px
-
-  .k-select__select
-   margin-left: 5px
-   width: 130px
-   max-width: 130px
-   height: 100%
 
 .k-select__title
  padding-bottom: 10px
@@ -148,5 +170,7 @@ export default {
 
  &:hover *
   color: dark-theme-variables.$fill-shortkey
-  background-color: dark-theme-variables.$fill-button-bg-h
+
+.k-select__select option.not_matched
+ display: none
 </style>